Extract all common logic of livesync services to platform-livesync-service-base
